requirements.txt
tensorflow
tensorflow-addons
numpy
scikit-image
matplotlib
imgaug
Pillow
tqdm